refactor(demo): moved demo to lib
