+++ src/network.ts
@@ -3,6 +3,7 @@ import { ref, watch } from '@vue/composition-api';
 import { SignedTransaction } from '@nimiq/hub-api';
 import type { Client, PlainStakingContract, PlainTransactionDetails } from '@nimiq/core';
 
+import { checkIfDemoIsActive } from '@/lib/Demo';
 import { useAddressStore } from './stores/Address';
 import { useTransactionsStore, TransactionState } from './stores/Transactions';
 import { useNetworkStore } from './stores/Network';
@@ -12,7 +13,6 @@ import { AddStakeEvent, ApiValidator, RawValidator, useStakingStore } from './st
 import { ENV_MAIN, STAKING_CONTRACT_ADDRESS } from './lib/Constants';
 import { reportToSentry } from './lib/Sentry';
 import { useAccountStore } from './stores/Account';
-import { useDemoStore } from './stores/Demo';
 
 let isLaunched = false;
 let clientPromise: Promise<Client>;
@@ -103,7 +103,6 @@ export async function launchNetwork() {
     const transactionsStore = useTransactionsStore();
     const addressStore = useAddressStore();
     const stakingStore = useStakingStore();
-    const demoStore = useDemoStore();
 
     const subscribedAddresses = new Set<string>();
 
@@ -120,7 +119,7 @@ export async function launchNetwork() {
     network$.fetchingTxHistory--;
 
     async function updateBalances(addresses: string[] = [...balances.keys()]) {
-        if (!addresses.length || demoStore.isDemoEnabled) return;
+        if (!addresses.length || checkIfDemoIsActive()) return;
         await client.waitForConsensusEstablished();
         const accounts = await retry(() => client.getAccounts(addresses)).catch(reportFor('getAccounts'));
         if (!accounts) return;
@@ -305,7 +304,7 @@ export async function launchNetwork() {
     })();
 
     function transactionListener(plain: PlainTransactionDetails) {
-        if (demoStore.isDemoEnabled) return;
+        if (checkIfDemoIsActive()) return;
         if (plain.recipient === STAKING_CONTRACT_ADDRESS) {
             if (plain.data.type === 'add-stake') {
                 if (!balances.has(plain.sender) && 'staker' in plain.data) {
@@ -343,7 +342,7 @@ export async function launchNetwork() {
     }
 
     function subscribe(addresses: string[]) {
-        if (demoStore.isDemoEnabled) return false;
+        if (checkIfDemoIsActive()) return false;
         client.addTransactionListener(transactionListener, addresses);
         updateBalances(addresses);
         updateStakes(addresses);
@@ -353,7 +352,7 @@ export async function launchNetwork() {
     // Subscribe to new addresses (for balance updates and transactions)
     // Also remove logged out addresses from fetched (so that they get fetched on next login)
     watch(addressStore.addressInfos, () => {
-        if (demoStore.isDemoEnabled) return;
+        if (checkIfDemoIsActive()) return;
         const newAddresses: string[] = [];
         const removedAddresses = new Set(subscribedAddresses);
 
@@ -385,7 +384,7 @@ export async function launchNetwork() {
 
     // Fetch transactions for active address
     watch([addressStore.activeAddress, txFetchTrigger], ([activeAddress, trigger]) => {
-        if (demoStore.isDemoEnabled) return;
+        if (checkIfDemoIsActive()) return;
         const address = activeAddress as string | null;
         if (!address || fetchedAddresses.value.includes(address)) return;
         addFetchedAddress(address);
@@ -434,7 +433,7 @@ export async function launchNetwork() {
     // Fetch transactions for proxies
     const proxyStore = useProxyStore();
     watch(proxyStore.networkTrigger, () => {
-        if (demoStore.isDemoEnabled) return;
+        if (checkIfDemoIsActive()) return;
         const newProxies: string[] = [];
         const addressesToSubscribe: string[] = [];
         for (const proxyAddress of proxyStore.allProxies.value) {
